// lib/models/new-user.ts
export default class NewUser {
  isBusiness: boolean = false;
  firstName: string = "";
  lastName: string = "";
  email: string = "";
  documentType: string = "";
  document: string = "";
  username: string = "";
  phoneNumber: string = "";
  businessName: string = "";
  tradename: string = "";
  address: string = "";
  password: string = "";
  repeatPassword: string = "";
}